+++ src/Jchemo.jl
@@ -424,7 +424,7 @@ export
     transfbl, 
     ## Validation
     residreg, residcla, 
-    ssr, msep, rmsep, rmsepstand, rrmsep,
+    ssr, msep, rmsep, rmsepstand, rrmsep, mae,
     bias, sep, cor2, r2, rpd, rpdr, mse, 
     errp, merrp,
     mpar,