+++ docs/src/news.md
@@ -5,6 +5,7 @@
 - News
     - Function **aggmean**: Compute column-wise means by class in a dataset.
     - Function **finduniq**: Find the indexes making unique the IDs in a ID vector.
+    - Function **mae**: Compute the mean absolute error.
     - Function **rrmsep**: Compute the relative RMSEP.
 
 - Modifications